refact: moved more generators around
