+++ xgi/generators/random.py
@@ -5,22 +5,16 @@ import warnings
 from collections import defaultdict
 from itertools import combinations
 
-import networkx as nx
 import numpy as np
 from scipy.special import comb
 
-from ..classes import SimplicialComplex
 from .classic import empty_hypergraph
 from .lattice import ring_lattice
-from .simplicial_complexes import flag_complex_d2
 
 __all__ = [
     "chung_lu_hypergraph",
     "dcsbm_hypergraph",
     "random_hypergraph",
-    "random_simplicial_complex",
-    "random_flag_complex_d2",
-    "random_flag_complex",
     "watts_strogatz_hypergraph",
 ]
 
@@ -116,7 +110,8 @@ def chung_lu_hypergraph(k1, k2, seed=None):
 
 
 def dcsbm_hypergraph(k1, k2, g1, g2, omega, seed=None):
-    """A function to generate a DCSBM hypergraph.
+    """A function to generate a Degree-Corrected Stochastic Block Model
+    (DCSBM) hypergraph.
 
     Parameters
     ----------
@@ -311,147 +306,6 @@ def random_hypergraph(N, ps, seed=None):
     H.add_edges_from(hyperedges)
 
     return H
-
-
-def random_simplicial_complex(N, ps, seed=None):
-    """Generates a random hypergraph
-
-    Generate N nodes, and connect any d+1 nodes
-    by a simplex with probability ps[d-1]. For each simplex,
-    add all its subfaces if they do not already exist.
-
-    Parameters
-    ----------
-    N : int
-        Number of nodes
-    ps : list of float
-        List of probabilities (between 0 and 1) to create a
-        hyperedge at each order d between any d+1 nodes. For example,
-        ps[0] is the wiring probability of any edge (2 nodes), ps[1]
-        of any triangles (3 nodes).
-    seed : int or None (default)
-        The seed for the random number generator
-
-    Returns
-    -------
-    Simplicialcomplex object
-        The generated simplicial complex
-
-    References
-    ----------
-    Described as 'random simplicial complex' in
-    "Simplicial Models of Social Contagion", Nature Communications 10(1), 2485,
-    by I. Iacopini, G. Petri, A. Barrat & V. Latora (2019).
-    https://doi.org/10.1038/s41467-019-10431-6
-
-    Example
-    -------
-    >>> import xgi
-    >>> H = xgi.random_simplicial_complex(20, [0.1, 0.01])
-
-    """
-
-    if seed is not None:
-        np.random.seed(seed)
-
-    if (np.any(np.array(ps) < 0)) or (np.any(np.array(ps) > 1)):
-        raise ValueError("All elements of ps must be between 0 and 1 included.")
-
-    nodes = range(N)
-    simplices = []
-
-    for i, p in enumerate(ps):
-        d = i + 1  # order, ps[0] is prob of edges (d=1)
-
-        potential_simplices = combinations(nodes, d + 1)
-        n_comb = comb(N, d + 1, exact=True)
-        mask = np.random.random(size=n_comb) <= p  # True if simplex to keep
-
-        simplices_to_add = [e for e, val in zip(potential_simplices, mask) if val]
-
-        simplices += simplices_to_add
-
-    S = SimplicialComplex()
-    S.add_nodes_from(nodes)
-    S.add_simplices_from(simplices)
-
-    return S
-
-
-def random_flag_complex_d2(N, p, seed=None):
-    """Generate a maximal simplicial complex (up to order 2) from a
-    :math:`G_{N,p}` Erdős-Rényi random graph by filling all empty triangles with 2-simplices.
-
-    Parameters
-    ----------
-    N : int
-        Number of nodes
-    p : float
-        Probabilities (between 0 and 1) to create an edge
-        between any 2 nodes
-    seed : int or None (default)
-        The seed for the random number generator
-
-    Returns
-    -------
-    SimplicialComplex
-
-    Notes
-    -----
-    Computing all cliques quickly becomes heavy for large networks.
-    """
-    if seed is not None:
-        random.seed(seed)
-
-    if (p < 0) or (p > 1):
-        raise ValueError("p must be between 0 and 1 included.")
-
-    G = nx.fast_gnp_random_graph(N, p, seed=seed)
-
-    return flag_complex_d2(G)
-
-
-def random_flag_complex(N, p, max_order=2, seed=None):
-    """Generate a flag (or clique) complex from a
-    :math:`G_{N,p}` Erdős-Rényi random graph by filling all cliques up to dimension max_order.
-
-    Parameters
-    ----------
-    N : int
-        Number of nodes
-    p : float
-        Probability (between 0 and 1) to create an edge
-        between any 2 nodes
-    max_order : int
-        maximal dimension of simplices to add to the output simplicial complex
-    seed : int or None (default)
-        The seed for the random number generator
-
-    Returns
-    -------
-    SimplicialComplex
-
-    Notes
-    -----
-    Computing all cliques quickly becomes heavy for large networks.
-    """
-
-    if (p < 0) or (p > 1):
-        raise ValueError("p must be between 0 and 1 included.")
-
-    G = nx.fast_gnp_random_graph(N, p, seed=seed)
-
-    nodes = G.nodes()
-    edges = list(G.edges())
-
-    # compute all triangles to fill
-    max_cliques = list(nx.find_cliques(G))
-
-    S = SimplicialComplex()
-    S.add_nodes_from(nodes)
-    S.add_simplices_from(max_cliques, max_order=max_order)
-
-    return S
 
 
 def watts_strogatz_hypergraph(n, d, k, l, p, seed=None):
